+++ voto/viewmodels/shared/viewmodelbase.py
@@ -8,6 +8,7 @@ class ViewModelBase:
     def __init__(self):
         self.request: Request = flask.request
         self.request_dict = request_dict.create("")
+        self.error = None
 
     def to_dict(self):
         return self.__dict__
